inference/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
joblib>=1.0.2
markdown>=3.3.4
python-docx>=0.8.11 